requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
APScheduler==3.11.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.6.0
greenlet==3.2.1
http_ece==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
propcache==0.3.1
py-vapid==1.9.2
pycparser==2.22
python-dotenv==1.1.0
pytz==2025.2
pywebpush==2.0.3
redis==5.2.1
requests==2.32.3
six==1.17.0
spotipy==2.25.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.20.0
ytmusicapi
requests
Pillow
flask_mail
email_validator
better-profanity==0.7.0
